Server/requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.6
gunicorn==22.0.0
python-jose[cryptography]>=3.3.0
google-auth>=2.35.0
requests>=2.32.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.0
python-multipart>=0.0.12
pydantic>=2.9.2
pydantic-settings>=2.5.2
psycopg[binary,pool]>=3.2.3
python-dotenv
